+++ src/gromo/modules/growing_module.py
@@ -473,6 +473,7 @@ class GrowingModule(torch.nn.Module):
         tensor_s_shape: tuple[int, int] | None = None,
         tensor_m_shape: tuple[int, int] | None = None,
         post_layer_function: torch.nn.Module = torch.nn.Identity(),
+        extended_post_layer_function: torch.nn.Module | None = None,
         allow_growing: bool = True,
         previous_module: torch.nn.Module | None = None,
         next_module: torch.nn.Module | None = None,
@@ -530,6 +531,25 @@ class GrowingModule(torch.nn.Module):
 
         self.layer: torch.nn.Module = layer.to(self.device)
         self.post_layer_function: torch.nn.Module = post_layer_function.to(self.device)
+        if extended_post_layer_function is None:
+            self.extended_post_layer_function = self.post_layer_function
+        else:
+            self.extended_post_layer_function = extended_post_layer_function.to(
+                self.device
+            )
+        if isinstance(self.extended_post_layer_function, torch.nn.Sequential):
+            for module in self.extended_post_layer_function:
+                if hasattr(module, "num_features"):
+                    warnings.warn(
+                        f"Warning in {self.name}: The extended post layer "
+                        f"function may get a variable input size."
+                    )
+        elif hasattr(self.extended_post_layer_function, "num_features"):
+            warnings.warn(
+                f"Warning in {self.name}: The extended post layer "
+                f"function may get a variable input size."
+            )
+
         self._allow_growing = allow_growing
         assert not self._allow_growing or isinstance(
             previous_module, (GrowingModule, MergeGrowingModule)
@@ -820,7 +840,9 @@ class GrowingModule(torch.nn.Module):
             supplementary_pre_activity = (
                 self._scaling_factor_next_module * self.extended_output_layer(x)
             )
-            supplementary_activity = self.post_layer_function(supplementary_pre_activity)
+            supplementary_activity = self.extended_post_layer_function(
+                supplementary_pre_activity
+            )
         else:
             supplementary_activity = None
 
@@ -876,7 +898,7 @@ class GrowingModule(torch.nn.Module):
                 )
                 return self.next_module.input
         else:
-            raise ValueError("The pre-activity is not stored.")
+            raise ValueError(f"The pre-activity is not stored for {self.name}.")
 
     # Statistics computation
     def projected_v_goal(self, input_vector: torch.Tensor) -> torch.Tensor:
@@ -1147,7 +1169,7 @@ class GrowingModule(torch.nn.Module):
         raise NotImplementedError
 
     def _apply_output_changes(
-        self, scaling_factor: float | torch.Tensor | None = None
+        self, scaling_factor: float | torch.Tensor | None = None, extension_size: int = 0
     ) -> None:
         """
         Extend the layer output with the current layer output extension,
@@ -1180,12 +1202,20 @@ class GrowingModule(torch.nn.Module):
             ),
         )
 
+        if isinstance(self.post_layer_function, torch.nn.Sequential):
+            for module in self.post_layer_function:
+                if hasattr(module, "grow"):
+                    module.grow(extension_size)
+        elif hasattr(self.post_layer_function, "grow"):
+            self.post_layer_function.grow(extension_size)
+
     def apply_change(
         self,
         scaling_factor: float | torch.Tensor | None = None,
         apply_previous: bool = True,
         apply_delta: bool = True,
         apply_extension: bool = True,
+        extension_size: int | None = None,
     ) -> None:
         """
         Apply the optimal delta and extend the layer with current
@@ -1206,6 +1236,9 @@ class GrowingModule(torch.nn.Module):
             if True apply the optimal delta to the layer, by default True
         apply_extension: bool
             if True apply the extension to the layer, by default True
+        extension_size: int | None
+            size of the extension to apply, by default None and get automatically
+            determined using `self.eigenvalues_extension.shape[0]`
         """
         # print(f"==================== Applying change to {self.name} ====================")
         if scaling_factor is not None:
@@ -1237,9 +1270,18 @@ class GrowingModule(torch.nn.Module):
                 )
 
             if apply_previous and self.previous_module is not None:
+                if extension_size is None:
+                    assert self.eigenvalues_extension is not None, (
+                        "We need to determine the size of the extension but"
+                        "it was not given as parameter nor could be automatically"
+                        "determined as self.eigenvalues_extension is None"
+                        f"(Error occurred in {self.name})"
+                    )
+                    extension_size = self.eigenvalues_extension.shape[0]
                 if isinstance(self.previous_module, GrowingModule):
                     self.previous_module._apply_output_changes(
-                        scaling_factor=self.scaling_factor
+                        scaling_factor=self.scaling_factor,
+                        extension_size=extension_size,
                     )
                 elif isinstance(self.previous_module, MergeGrowingModule):
                     raise NotImplementedError  # TODO